requirements.txt
python-telegram-bot==20.8
web3==6.15.1
requests==2.31.0
python-dotenv==1.0.1
aiohttp==3.9.3
asyncio==3.4.3 